requirements.txt
pandas
torch
torchvision
torchinfo
numpy
pillow
